feat: add request logging for stock updates
