+++ woocommerce_fusion/woocommerce/doctype/woocommerce_order/test_woocommerce_order.py
@@ -9,10 +9,10 @@ from urllib.parse import urlparse
 import frappe
 from frappe.tests.utils import FrappeTestCase
 
+from woocommerce_fusion.tasks.utils import APIWithRequestLogging
 from woocommerce_fusion.woocommerce.doctype.woocommerce_order.woocommerce_order import (
 	API,
 	WC_ORDER_DELIMITER,
-	APIWithRequestLogging,
 	WooCommerceAPI,
 	WooCommerceOrder,
 	generate_woocommerce_order_name_from_domain_and_id,
